History Services. *Changes: -Added new table <Deck> ( contains cards for play). -Now table <Card> contains cards, that used in game. -Adds UserID to the <PlayerHand>. *Services: -GameByUserID ( with steps played ).

+++ BlackJack.DataAccess/Repositories/GameRepository.cs
@@ -18,7 +18,7 @@ namespace BlackJack.DataAccess.Repositories
 
         }
 
-        public IEnumerable<Game> SelectGamesByUserId(int userId)
+        public IEnumerable<Game> SelectGamesByUserId(long userId)
         {
             List<Game> games = _dbSet.Where(e => e.UserID == userId).ToList();
